The feed URLs that were being used didn't always work for all users (notably unlike the previous code had, the value of 'sf' seems to vary from user to user.)
There are probably better ways of doing this then what I've got here, consider this a bit of a hack.  If anyone else gets some free time, feel free to go over it in greater detail.

I've also added functionality for iTunes' 'Favourite Artists' feed which I was at it.

Signed-off-by: Alan Isherwood (AlanI) <alan@alanisherwood.id.au>